use PROJECT_NAME for all user-facing strings in GUI
This most likely break a lot of translations. The .po files need to be
updated accordingly to replace all occurences of "Electron Cash"
by "Electrum BCHA".
Better to wait a bit for the branding to be confirmed before doing that.

Test Plan:

Run the GUI and check that occurences of "Electron Cash" have been
replaced by "Electrum BCHA"